@@ -12,7 +12,7 @@
         <AssemblyName>ModSetting</AssemblyName>
         <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
         <FileAlignment>512</FileAlignment>
-        <LangVersion>8</LangVersion>
+        <LangVersion>9</LangVersion>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
         <PlatformTarget>AnyCPU</PlatformTarget>
@@ -642,6 +642,7 @@
         <Compile Include="ModBehaviour.cs" />
         <Compile Include="Properties\AssemblyInfo.cs"/>
         <Compile Include="Saver.cs" />
+        <Compile Include="UI\ButtonUI.cs" />
         <Compile Include="UI\DropDownUI.cs" />
         <Compile Include="UI\GlobalPanelUI.cs" />
         <Compile Include="UI\InputUI.cs" />
